apps/processes/serializers.py
from rest_framework import serializers
from django.contrib.auth.models import User


class RequestSerializer(serializers.Serializer):
    email = serializers.EmailField()
    message = serializers.CharField(max_length=200)
    
    


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['username', 'email', 'password']
        extra_kwargs = {
            'password': {'write_only': True}
        }

    def create(self, validated_data):
        user = User.objects.create_user(
            username=validated_data['username'],
            email=validated_data['email'],
            password=validated_data['password']
        )
        return user



class LoginSerializer(serializers.Serializer):
    username = serializers.CharField(max_length=150)
    password = serializers.CharField(max_length=128, write_only=True)