requirements.txt
Django==5.2
djangorestframework==3.16.0
celery==5.5.1
redis==5.2.1
drf-yasg==1.21.10
python-decouple==3.8
psycopg2-binary==2.9.10